finish dataclass support and pull out type conversion utilities
